api-microservice/src/main/java/io/github/brunogabriel/apimicroservice/service/PetService.java
package io.github.brunogabriel.apimicroservice.service;

import io.github.brunogabriel.apimicroservice.domain.Pet;
import io.github.brunogabriel.apimicroservice.repository.PetRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class PetService {
    @Autowired
    private PetRepository petRepository;

    public List<Pet> findAllByClientId(String clientId) {
        return petRepository.findAllByClientId(clientId);
    }
}
